Trait xcm_transactor::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>;
    type CurrencyId: Parameter + Member + Clone;
    type CurrencyIdToMultiLocation: Convert<Self::CurrencyId, Option<MultiLocation>>;
    type Transactor: Parameter + Member + Clone + XcmTransact;
    type AssetTransactor: TransactAsset;
    type DerivativeAddressRegistrationOrigin: EnsureOrigin<Self::Origin>;
    type AccountIdToMultiLocation: Convert<Self::AccountId, MultiLocation>;
    type Weigher: WeightBounds<Self::Call>;
    type LocationInverter: InvertLocation;
    type SelfLocation: Get<MultiLocation>;
    type SovereignAccountDispatcherOrigin: EnsureOrigin<Self::Origin>;
    type XcmSender: SendXcm;
    type BaseXcmWeight: Get<Weight>;
    type ReserveProvider: Reserve;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>

The balance type.

type CurrencyId: Parameter + Member + Clone

Currency Id.

type CurrencyIdToMultiLocation: Convert<Self::CurrencyId, Option<MultiLocation>>

Convert T::CurrencyId to MultiLocation.

type Transactor: Parameter + Member + Clone + XcmTransact

type AssetTransactor: TransactAsset

AssetTransactor allows us to withdraw asset without being trapped This should change in xcm v3, which allows us to burn assets

type DerivativeAddressRegistrationOrigin: EnsureOrigin<Self::Origin>

type AccountIdToMultiLocation: Convert<Self::AccountId, MultiLocation>

Convert T::AccountId to MultiLocation.

type Weigher: WeightBounds<Self::Call>

Means of measuring the weight consumed by an XCM message locally.

type LocationInverter: InvertLocation

Means of inverting a location.

type SelfLocation: Get<MultiLocation>

Self chain location.

type SovereignAccountDispatcherOrigin: EnsureOrigin<Self::Origin>

type XcmSender: SendXcm

XCM sender.

type BaseXcmWeight: Get<Weight>

The actual weight for an XCM message is T::BaseXcmWeight + T::Weigher::weight(&msg).

type ReserveProvider: Reserve

The way to retrieve the reserve of a MultiAsset. This can be configured to accept absolute or relative paths for self tokens

type WeightInfo: WeightInfo